+++ spec/grumlin/repository_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 RSpec.describe Grumlin::Repository, gremlin_server: true do
-  let(:repository_klass) do
+  let(:repository_class) do
     Class.new do
       extend Grumlin::Repository
 
@@ -10,12 +10,12 @@ RSpec.describe Grumlin::Repository, gremlin_server: true do
       end
     end
   end
-  let(:repository) { repository_klass.new }
+  let(:repository) { repository_class.new }
 
   describe "class methods" do
     %i[shortcut shortcuts shortcuts_from].each do |method|
       it "responds to ##{method}" do
-        expect(repository_klass).to respond_to(method)
+        expect(repository_class).to respond_to(method)
       end
     end
   end
@@ -30,7 +30,143 @@ RSpec.describe Grumlin::Repository, gremlin_server: true do
 
   describe "included shortcuts" do
     it "includes shortcuts" do
-      expect(repository_klass.shortcuts.keys).to eq(%i[props hasAll shortcut])
+      expect(repository_class.shortcuts.keys).to eq(%i[props hasAll shortcut])
+    end
+  end
+
+  describe "extending" do
+    context "when not inherited" do
+      it "can be extended multiple times" do
+        expect { repository_class.extend(described_class) }.not_to raise_error
+      end
+    end
+
+    context "when inherited" do
+      it "can be extended again" do
+        expect do
+          Class.new(repository_class) do
+            extend Grumlin::Repository
+          end
+        end.not_to raise_error
+      end
+    end
+  end
+
+  describe "inheritance" do
+    describe "when inherited" do
+      let(:child) { Class.new(repository_class) }
+
+      it "injects it's shortcuts to the child class" do
+        expect(child.new.g).to respond_to(:shortcut)
+      end
+    end
+  end
+
+  describe "importing shortcuts from a module" do
+    subject { repository_class.shortcuts_from(shortcut_module) }
+
+    let(:shortcut_module) do
+      Module.new do
+        extend Grumlin::Shortcuts
+
+        shortcut :another_shortcut do
+          property(:another_shortcut, true)
+        end
+      end
+    end
+
+    context "when there are no shortcut naming conflicts" do
+      it "imports shortcuts" do
+        subject
+        expect(repository.g).to respond_to(:another_shortcut)
+      end
+    end
+
+    context "when there is a naming conflict" do
+      context "when conflicting shortcuts point to one implementation" do
+        before do
+          repository_class.shortcuts_from(shortcut_module)
+        end
+
+        it "successfully imports shortcuts" do
+          expect { subject }.not_to raise_error
+        end
+      end
+
+      context "when conflicting shortcuts point to different implementations" do
+        let(:another_shortcut_module) do
+          Module.new do
+            extend Grumlin::Shortcuts
+
+            shortcut :another_shortcut do
+              property(:another_shortcut, true)
+            end
+          end
+        end
+
+        before do
+          repository_class.shortcuts_from(another_shortcut_module)
+        end
+
+        include_examples "raises an exception", ArgumentError, "shortcut 'another_shortcut' already exists"
+      end
+    end
+  end
+
+  describe "importing shortcuts from another repository" do
+    subject { repository_class.shortcuts_from(another_repository_class) }
+
+    let(:another_repository_class) do
+      Class.new do
+        extend Grumlin::Repository
+
+        shortcut :another_shortcut do
+          property(:another_shortcut, true)
+        end
+      end
+    end
+
+    context "when there are no shortcut naming conflicts" do
+      it "imports shortcuts" do
+        subject
+        expect(repository.g).to respond_to(:another_shortcut)
+      end
+    end
+
+    context "when there is a naming conflict" do
+      context "when conflicting shortcuts point to one implementation" do
+        let(:another_repository_class) do
+          Class.new do # props and hasAll in this case
+            extend Grumlin::Repository
+          end
+        end
+
+        before do
+          repository_class.shortcuts_from(another_repository_class)
+        end
+
+        it "successfully imports shortcuts" do
+          expect { subject }.not_to raise_error
+        end
+      end
+
+      context "when conflicting shortcuts point to different implementations" do
+        let(:yet_another_repository_class) do
+          Class.new do
+            extend Grumlin::Repository
+
+            shortcut :another_shortcut do
+              property(:another_shortcut, true)
+            end
+          end
+        end
+
+        before do
+          repository_class.shortcuts_from(yet_another_repository_class)
+        end
+
+        include_examples "raises an exception", ArgumentError, "shortcut 'another_shortcut' already exists"
+      end
     end
   end
 end